bitcoind: split polling logic in a new submodule
Achieved by moving the "main" of the bitcoind thread at the root of the module, and
renaming actions.rs to poller.rs where most of the logic was the poller's anyways.
